IGN-6250: Support unsigned modules in gradle-module-plugin (#30)
* IGN-6250: Support unsigned modules in gradle-module-plugin
This adds support to skip the module signing requirement and task for the gradle-module-plugin. This adds the `skipModlSigning` boolean property to the ModuleSetting extension. This property doesn't disrupt task ordering or execution, it merely alters the assignment of task input and outputs to assign the unsigned module file and skips the `signModule` task. It also adds some additional test cases to test that the `buildResult.json`'s filename points to the unsigned and that the signed module is not built if the prop is true.