IGN-6325: `collectModlDependencies` detects source changes
A subproject with a new, deleted, or mutated source puts the
`collectModlDependencies` task in an out-of-date state so that it
will re-fire and pull the new subproject JAR into the module's build
artifacts directory.

The source change detection is indirect; only source changes marking
a subproject's `jar` task as out-of-date will in turn mark
`collectModlDependencies` as out-of-date. By default that means
changes to `src/main/<lang>/*` and `src/main/resources/*` sources.
If you have reconfigured the subproject's `jar` task such that other
sources serves as inputs to it, those source changes will inform
`collectModlDependencies` as well.