IGN-7871 PKCS#11 signing
Add HSM/hardware token keystore capability via new property
`pkcs11CfgFile`. Unlike filesystem-based `keystoreFile`, the signing
key is stored in a device such as a YubiKey, with driver and other
`SunPKCS11` provider properties specified in a configuration file for
the `signModule` task.

Other changes include refactoring existing tests in `SignModuleTest`,
Spotless delinting in `gradle-module-plugin`, and a few minor test
and typo fixes.

NOTE 1: This is currently versioned at `0.2.0-SNAPSHOT`. After the PR
is approved and any other validation--preferably some some testing by
IA and partner developers--this will bump to `0.2.0`.

NOTE 2: This was integration tested against a YubiKey 5 NFC with the
wrinkle that the slot `9a` authorization key was used, not the `9c`
signing key. There are notes in `SignModuleTest` as to why this
workaround is necessary. (At some point we might be able to sign with
the "correct" YK5 slot but it appears to require some tricky callback
code.)

NOTE 3: Because we call down to `module-signer` to do the actual
signing, and because its filter is narrowed to keys of type
`SHA256withRSA`, the key in the keystore must be that type. (This
constraint predates the current enhancement.) At some point we'll
likely enhance `module-signer` to support more key algorithms such
as Elliptic Curve and whatever post-quantum algorithms ultimately pass
muster with NIST and related entities.