IGN-9137 `moduleDependencySpecs` impl change
Switch to more Gradle-ey `NamedDomainObjectContainer` implementation.
That buys us a few things: 1) less code to implement our DSL, 2) syntax
between Groovy and Kotlin build script DSL is identical in most use
cases, and very close in others, 3) allows for super-compact "magic"
syntax demonstrated in one of the new unit tests.

This implementation also removes some redundant and potentially error-
causing defaulting between from old to new project extension
properties. It's largely covered by the code we already have in
`WriteModuleXml.buildXml`, and we could probably tighten it up further
if we need to.

Last, by leaning on native Gradle DSL APIs we don't have to rename
the helper function `ModuleSettings.moduleId`, because we don't need it
anymore. Otherwise we probably would've to renamed it `module` or
similar, because `moduleId` is also the name of the uniquely
identifying field/property.