Dev window version prep
* Generator core and CLI: `0.1.0-SNAPSHOT` > `0.4.0-SNAPSHOT`.
* Module plugin: `0.3.0` > `0.4.0-SNAPSHOT`.

Generator currently only scoped for bug fixes, but syncing its version
with that of the module plugin for ease of maintenance. Module
plugin has bug fixes and an improvement, hence the minor
version bump.

`generator` and `gradle-module-plugin` are included builds, not
subprojects, so there's not an easy way to just set a single version
from the root build script.

With fixes for dangling old version strings throughout, especially for
generator sources referencing the module plugin version.

These will become `0.4.0` before we release.