IGN-10612 feedback from IA module devs
Add `--foldJars` task option to preserve the old behavior by default,
it's a breaking change and maybe the dups are useful to some.

The option if passed does the scope folding and the sorting as before
and another sort of folding for `CDG`-scoped jars, turning them into
the `A`[ll]-scoped equivalent.

Plus noticed that the tests fall into the `assertEquals` API trap where
you think the argument order is `actual, expected` but it's really
`expected, actual`. Completely nonintuitive and sort of asymmetric to
other `assert*` APIs but when `assertEquals` fails having them in the
wrong order makes things harder to troubleshoot.